I enjoyed getting my review back, and am implementing the changes that were given. Thank you so much for taking the time to review mu code, reviewer person. https://review.udacity.com/#!/reviews/1674764
Review Application: Reformatted code via Ctrl + Alt + L in Android Studio so it complys with XML and Java readability conventions.

I don't like having the XML ending at the end of an attribute; I'd rather have each attribute on a separate line in such a way that I can move them around and it does not impact how the function works. With that ending on the last attribute, care must be taken to not move the ending when moving the last attribute.
I have the same issue with this line at the top: <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android". I would rather the attribute be on a separate line.

Regardless, I want my code to comply with global conventions, so if that is the standard I will do my best to get used to it.